python-src/requirements.txt
mcp>=1.0.0
crewai>=0.22.0
sqlalchemy>=2.0.0
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
cryptography>=41.0.0